implemented feedback, made gradebook download a util function and updated downloading logic
