+++ devU-client/src/assets/global.scss
@@ -53,7 +53,7 @@
     margin-bottom: 20px;
   }
 
-  button.btnPrimary {
+  .btnPrimary {
     @extend .btn;
     background-color: var(--primary);
     border: 3px solid var(--primary);
@@ -65,7 +65,7 @@
     }
   }
 
-  button.btnSecondary {
+  .btnSecondary {
     @extend .btn;
 
     background-color: var(--btn-secondary-background);
@@ -77,7 +77,7 @@
     }
   }
 
-  button.btnDelete {
+  .btnDelete {
     @extend .btn;
     background-color: var(--btn-delete-background);
     color: var(--btn-delete-text);
