+++ devU-client/src/components/pages/forms/assignments/textProblemModal.tsx
@@ -68,7 +68,6 @@ const TextProblemModal = ({ open, onClose }: Props) => {
         await RequestService.post(`/api/course/${courseId}/assignment/${assignmentId}/non-container-auto-graders/`, graderFormData)
             .then(() => {
                 console.log("GRADER CREATED")
-                //window.location.reload()
             })
             .catch((err: ExpressValidationError[] | Error) => {
                 const message = Array.isArray(err) ? err.map((e) => `${e.param} ${e.msg}`).join(', ') : err.message
