+++ devU-client/src/components/pages/gradebook/gradebookStudentPage.scss
@@ -16,6 +16,11 @@
     scrollbar-color: $primary;
   }
 
+  .download{
+    float: right;
+    text-decoration: none;
+}
+
   table {
         border-radius: 20px;
         gap:10px;
